pyproject.toml
[project]
authors = [{name = "Hoog-v", email = "hogeweyv@gmail.com"}]
name = "database_builder_scepa"
requires-python = "== 3.11"
version = "0.1.0"
dependencies = [
    "typedb-driver==2.29.2",
    "fastapi>=0.128.0",
    "pyzotero>=1.7.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "anyio>=4.12.1",
    "spacy>=3.8.11",
    "loguru>=0.7.3",
    "torch",
    "torchvision",
    "testcontainers",
    "qdrant-client"
, "docling>=2.73.0,<3", "database-builder-libs @ git+https://github.com/AIM-kennisplatformen/database-builder-libs.git", "span-marker>=1.7.0,<2", "openai>=2.24.0,<3", "pypdf>=6.7.3,<7"]

[dependency-groups]
dev = [
  "lefthook",
  "debugpy>=1.8.19",
  "httpretty>=1.1.4",
  "pandas-stubs>=2.3.3.260113",
  "pytest>=9.0.2",
  "ruff>=0.14.11",
  "ty>=0.0.12",
  "types-openpyxl>=3.1.5.20250919",
  "types-requests>=2.32.4.20260107",
  "lingua-language-detector"
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/backend", "src/database_builder_libs"]

[tool.pixi.workspace]
channels = ["https://fast.prefix.dev/conda-forge",
  "https://repo.prefix.dev/pypdfium2-public",
  "https://repo.prefix.dev/pytorch"]
platforms = ["linux-64","osx-arm64", "win-64"]

[tool.pixi.dependencies]
# Dotenv is for reading .env files
compilers = ">=1.9.0"
pkg-config = ">=0.29.2"


[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = ["--import-mode=importlib"]
testpaths = ["tests/backend"]
filterwarnings = [
    "ignore:datetime.datetime.utcnow.*:DeprecationWarning"
]

[tool.ty.src]
include=["src"]
exclude=["tests/", ".pixi/"]

[tool.ruff.lint]
select = ["B024", "E4", "E7", "E9", "F", "C4", "PT"]

[tool.ruff]
exclude = [".pixi", "tests"]

[tool.pixi.tasks]
lint = "ruff check"
format = "ruff format"
typecheck = "ty check ."
test = "pytest -s"
app = "python -m src.scepa_app.main"

[tool.pixi.activation.env]
PYTHONNOUSERSITE = "1"
LD_LIBRARY_PATH = "$CONDA_PREFIX/lib:$LD_LIBRARY_PATH"

# Platform-specific deps
[tool.pixi.target.linux-64.dependencies]
python-devtools = ">=0.12.2"


[tool.pixi.environments]
default = { features = ["dev"], solve-group = "default" }
development = { features = ["dev"], solve-group = "default" }

